Add more main-thread overrides to satisfy CLAP thread-check
These are all based on real-world situations where the host is using a
different ID than the UI thread ID.

VST3 documents these functions as [UI-thread]. We have to assume that
the host is taking the 'logically consistent' approach to the UI thread.
That is, that the thread ID can change so long as the plugin
simultaneous calls to main-thread functions at the same time.